# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LOOP API
  version: 1.0.0
  description: Tenant-scoped customer feedback intelligence API.
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /feedback:
    get:
      tags: [Feedback]
      summary: Search tenant feedback
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: q, in: query, schema: { type: string, maxLength: 200 } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/FeedbackStatus' } }
        - { name: sentiment, in: query, schema: { $ref: '#/components/schemas/Sentiment' } }
        - { name: channel, in: query, schema: { type: string } }
        - { name: themeId, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: sort, in: query, schema: { type: string, enum: [createdAt, updatedAt] }, required: false }
        - { name: direction, in: query, schema: { type: string, enum: [asc, desc] }, required: false }
      responses:
        '200': { description: Paginated feedback }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Feedback]
      summary: Create feedback and enqueue analysis
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeedbackCreate' }
      responses:
        '201': { description: Feedback accepted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /reports:
    get:
      tags: [Reports]
      summary: List reports
      responses:
        '200': { description: Reports returned }
    post:
      tags: [Reports]
      summary: Queue a Voice-of-Customer report
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, periodStart, periodEnd]
              properties:
                title: { type: string, maxLength: 120 }
                periodStart: { type: string, format: date-time }
                periodEnd: { type: string, format: date-time }
      responses:
        '202': { description: Report queued }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
...
  schemas:
...
    FeedbackCreate:
      type: object
      required: [content, channel]
      properties:
        content: { type: string, minLength: 1, maxLength: 20000 }
        channel: { type: string, minLength: 1, maxLength: 80 }
        customerLabel: { type: string, maxLength: 160 }
        sourceType: { type: string, enum: [MANUAL, CSV, SIMULATED, API, WEBHOOK] }
        externalId: { type: string, maxLength: 200 }
...
    FeedbackStatus:
      type: string
      enum: [NEW, REVIEWED, ACTIONED]
    Sentiment:
      type: string
      enum: [POS, NEU, NEG]
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
    Problem:
      type: object
      properties:
        error: { $ref: '#/components/schemas/Error' }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unauthorized:
      description: Authentication required
    Forbidden:
      description: Insufficient permission or cross-tenant access
...